Update description & name
